agent/requirements.txt
# Core dependencies for AgentCore Runtime Agent
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
bedrock-agentcore>=1.0.0
mcp>=1.0.0
boto3>=1.42.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiohttp>=3.8.0
nest_asyncio>=1.6.0
playwright>=1.40.0
